mcp-agent-experiment/pyproject.toml
[project]
name = "mcp-agent-experiment"
version = "0.1.0"
description = "MCP SQL Chatbot & Dashboard with PostgreSQL support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=1.2.6",
    "anthropic>=0.49.0",
    "python-dotenv>=0.9.9",
    "groq>=0.20.0",
    "mcp>=1.6.0",
    "openai>=1.70.0",
    "streamlit>=1.44.0",
    "asyncpg",
    "aiohttp",
    "google-genai>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
